split formLabel tests and added snapshot to other tests
