+++ src/Forms/FormItem.test.js
@@ -1,13 +1,62 @@
+import FormInput from './FormInput';
 import FormItem from './FormItem';
+import FormLabel from './FormLabel';
+import FormMessage from './FormMessage';
+import FormSelect from './FormSelect';
+import FormTextarea from './FormTextarea';
 import { mount } from 'enzyme';
 import React from 'react';
+import renderer from 'react-test-renderer';
 
-describe('FormItem Prop spreading', () => {
-    test('should allow props to be spread to the FormItem component', () => {
-        const element = mount(<FormItem data-sample='Sample' />);
+describe('<FormItem />', () => {
+    const formItem = (
+        <FormItem>
+            <FormLabel forAttr='input-1' required>
+                Default Input
+            </FormLabel>
+            <FormInput
+                id='input-1'
+                placeholder='Field placeholder text'
+                state='warning'
+                type='text' />
+            <FormTextarea className='blue' id='textarea-1'>
+                Pellentesque metus lacus commodo eget justo ut rutrum varius nunc.
+            </FormTextarea>
+            <FormTextarea id='textarea-2'>
+                Pellentesque metus lacus commodo eget justo ut rutrum varius nunc.
+            </FormTextarea>
+            <FormMessage type='help'>
+                Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
+            </FormMessage>
+            <FormMessage className='blue'>
+                Pellentesque metus lacus commodo eget justo ut rutrum varius nunc
+            </FormMessage>
+            <FormSelect id='select-1'>
+                <option value='1'>Duis malesuada odio volutpat elementum</option>
+                <option value='2'>Suspendisse ante ligula</option>
+                <option value='3'>Sed bibendum sapien at posuere interdum</option>
+            </FormSelect>
+            <FormSelect className='blue' disabled
+                id='select-1'>
+                <option value='1'>Duis malesuada odio volutpat elementum</option>
+            </FormSelect>
+        </FormItem>
+    );
 
-        expect(
-            element.getDOMNode().attributes['data-sample'].value
-        ).toBe('Sample');
+    test('create form item', () => {
+        // create form set with form inputs
+        let component = renderer.create(formItem);
+        let tree = component.toJSON();
+        expect(tree).toMatchSnapshot();
+    });
+
+    describe('Prop spreading', () => {
+        test('should allow props to be spread to the FormItem component', () => {
+            const element = mount(<FormItem data-sample='Sample' />);
+
+            expect(
+                element.getDOMNode().attributes['data-sample'].value
+            ).toBe('Sample');
+        });
     });
 });
